# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Booking App",
        "description": "This is the Event Booking App - API Documentation generated using Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "EventApp API Server"
        }
    ],
    "paths": {
...
        "/api/events": {
            "post": {
                "tags": [
                    "Events - Event Management"
                ],
                "summary": "List all events with pagination and state filter",
                "description": "Get a list of all events with pagination and event state filtering support.",
                "operationId": "listEvents",
                "requestBody": {
                    "description": "Pagination and event state filter data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "page_number"
                                ],
                                "properties": {
                                    "page_number": {
                                        "description": "The page number to fetch",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "location_country": {
                                        "description": "The location to filter events by country",
                                        "type": "string",
                                        "example": "England"
                                    },
                                    "event_state": {
                                        "description": "State of the event (upcoming, past)",
                                        "type": "string",
                                        "example": "upcoming"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of events",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event list fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "events": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Demo Event"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "This is demo event description"
                                                            },
                                                            "event_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2025-12-25"
                                                            },
                                                            "location_country": {
                                                                "type": "string",
                                                                "example": "England"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "total_pages": {
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/event/{id}": {
            "get": {
                "tags": [
                    "Events - Event Management"
                ],
                "summary": "View event details along with total bookings",
                "description": "Get the details of a specific event, including the event name, description, date, location, available seats, and total bookings.",
                "operationId": "viewEvent",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the event to fetch",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event details fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Event details fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Demo Event"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "This is demo event description"
                                                },
                                                "event_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2025-12-25"
                                                },
                                                "event_location_country": {
                                                    "type": "string",
                                                    "example": "England"
                                                },
                                                "seats_available": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "seats_booked": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Event not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Event not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}